Add initial PDF support
This uses the poppler commandline tool pdftotext. While it may be more
elegant and efficient to use a proper API, the commandline tool is more
widely available and avoids a dependency on an uncommon python package.
Furthermore, it only gets called for PDF files, so the performance
impact is limited.

Bug: